requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
orjson>=3.10.0
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
arcgis>=2.3.0
cachetools>=5.3.0
pydantic>=2.0.0
pandas>=2.0.0
